feat/job: build and upload windows cache
A job that generates a cache (a target directory) for use with Windows
builds. Due to the fact that Jenkins creates a new workspace for every
PR that's submitted, and due to the way Cargo is designed to rebuild
all dependencies in a 'local' target directory, this build will generate
a cache that each Windows build can pull in which will have all the
dependencies pre-built.

This also renames the safe-nd pipeline to use an underscore as a word
separator (even although that name is different from the actual repo
name).