refactor/job: remove image tag param
Since we're no longer versioning the build container along with the
application, and just using a build tag.

This may need to be changed at some point to possibly use the name of a
branch as a tag, but it'll be easy enough to change if that scenario
comes up.