style/job: consistent use of quotes
Mixed use of single or double quotes is annoying. This applies double
quotes consistently. The choice of single or double is usually quite
abitrary - the important thing is they're applied consistently. In the
case of Groovy, it doesn't do string interpolation for single quoted
strings, so the guide now is to use double quotes unless for some reason
you don't need string interpolation, in which case use single.