refactor/job: use build number in ami name
Use the build number as opposed to the commit hash, because the commit
hash isn't always changing when this build is run. We could do it a
different way and check that there is already an AMI with that commit
hash and do nothing, but we actually want a build to run sometimes even
if that's the case: for example if there's been a Rust upgrade, we need
the process to run again to pull in the new 'latest' tag.

The SAFE_COMMIT_HASH variable should really be renamed. I'll do that as
a separate PR on the safe-build-infrastructure repo.